no había visto que había un archivo.js para hacer la función de los filtros así que corte y pegue el codigo y agregué el sript correspondiente en index
