+++ index.html
@@ -78,6 +78,7 @@
     <script src="./src/menu.js" type="module"></script>
     <script src="./src/events.js" type="module"></script>
     <script src="./src/cart.js" type="module"></script>
+    <script src="./src/searcher.js" type="module"></script>
 </body>
 
 </html>
